refactor: HashMap -> EnumMap으로 변경

+++ src/main/java/lotto/view/ResultView.java
@@ -27,7 +27,7 @@ public class ResultView {
             .forEach(System.out::println);
     }
 
-    private static String getOneRankStatus(Rank rank, Integer winningCount) {
+    private static String getOneRankStatus(Rank rank, Long winningCount) {
         if (rank == Rank.SECOND) {
             return String.format("%d개 일치, 보너스 볼 일치(%d원)- %d개", rank.getValue(), rank.getMoney(), winningCount);
         }
